Changed Cat Movement
Changed the movement of the player from tank controls to cardinal directions.
Implementation looks a bit weird because I couldn't change the attributes of the class due to replication.
Also using Normalize on a vector made the cat disappear so I had to manually adjust the values of the vector to pseudo normalize it.